fix: 修复第三轮 CI 暴露的编译警告与冒烟脚本语法错误
- Linux（gcc -Werror=conversion）：abstractlogview.cpp 物理像素偏移
  计算 qint64 隐式转 qreal，改为显式 static_cast<double>。
- Mac（clang -Winconsistent-missing-override）：DocumentationWindow::
  reload 与 PredefinedFiltersComboBox::showPopup 补 override 标记
  （Qt 6.8 中 QTextBrowser::reload/QComboBox::showPopup 均为虚函数）。
- Update Smoke：验收步骤内联 PowerShell 的反引号续行在 CRLF 行尾下
  被 Windows PowerShell 5.1 误判（缺失字符串终止符），改单行调用。

本地 MSVC 增量构建 exit 0。

+++ src/ui/src/abstractlogview.cpp
@@ -2236,7 +2236,7 @@ void AbstractLogView::drawTextArea( QPaintDevice* paintDevice )
                           - static_cast<qint64>( firstLine_.get() );
     const auto visibleRows = static_cast<qint64>( getNbVisibleLines().get() );
     const auto totalRows = static_cast<qint64>( logData_->getNbLine().get() );
-    const double physicalDelta = rowDelta * charHeight_ * viewport()->devicePixelRatio();
+    const double physicalDelta = static_cast<double>( rowDelta ) * charHeight_ * viewport()->devicePixelRatio();
     const bool reuse = !useTextWrap_ && !textAreaCache_.invalid_
         && textAreaCache_.first_column_ == firstCol_
         && rowDelta != 0 && qAbs( rowDelta ) < visibleRows
